+++ include/php/eva-apiclient.php
@@ -158,7 +158,9 @@ $_eva_api_func = array(
                     'modinfo_lpi',
                     'modhelp_phi',
                     'modhelp_lpi',
-                    'assign_driver'
+                    'assign_driver',
+                    'set_phi_prop',
+                    'set_driver_prop'
                     ),
                 'cr' => array(
                     'update',
@@ -194,7 +196,9 @@ $_eva_api_func = array(
                     'get_driver',
                     'assign_driver',
                     'test_phi',
-                    'exec_phi'
+                    'exec_phi',
+                    'set_phi_prop',
+                    'set_driver_prop'
                     ),
                 'ce' => array(
                     'action',