+++ lib/eva/uc/driverapi.py
@@ -445,6 +445,18 @@ def load_driver(lpi_id, lpi_mod_id, phi_id, lpi_cfg=None, start=True):
     return lpi
 
 
+def set_phi_prop(phi_id, p, v):
+    phi = get_phi(phi_id)
+    cfg = phi.phi_cfg
+    phi_mod_id = phi.phi_mod_id
+    cfg[p] = v
+    if v is None: del cfg[p]
+    phi = load_phi(phi_id, phi_mod_id, cfg, start=True)
+    if phi:
+        phis[phi_id] = phi
+        return True
+
+
 def unload_phi(phi_id):
     phi = get_phi(phi_id)
     if phi is None: return False
@@ -470,6 +482,17 @@ def unload_phi(phi_id):
         eva.core.log_traceback()
     del phis[phi_id]
     return True
+
+
+def set_driver_prop(driver_id, p, v):
+    lpi = get_driver(driver_id)
+    cfg = lpi.lpi_cfg
+    cfg[p] = v
+    if v is None: del cfg[p]
+    lpi = load_driver(lpi.lpi_id, lpi.lpi_mod_id, lpi.phi_id, cfg, start=True)
+    if lpi:
+        drivers[driver_id] = lpi
+        return True
 
 
 def unload_driver(driver_id):